refactor: move close on outside click to modal wrapper
